requirements.txt
aioredis==1.3.1
asgiref==3.3.1
asttokens==2.0.4
async-timeout==3.0.1
attrs==20.3.0
autobahn==21.2.1
Automat==20.2.0
autopep8==1.5.5
Brotli==1.0.9
certifi==2020.12.5
cffi==1.14.5
channels==3.0.3
channels-redis==3.2.0
chardet==4.0.0
click==7.1.2
colorama==0.4.4
constantly==15.1.0
cryptography==3.4.5
cycler==0.10.0
daphne==3.0.1
dash==1.19.0
dash-core-components==1.15.0
dash-html-components==1.1.2
dash-renderer==1.9.0
dash-table==4.11.2
Django==3.1.6
django-plotly-dash==1.6.1
django-redis==4.12.1
django-rest-framework==0.1.0
djangorestframework==3.12.2
dpd-components==0.1.0
executing==0.5.4
fastapi==0.63.0
Flask==1.1.2
Flask-Compress==1.8.0
future==0.18.2
h11==0.12.0
hiredis==1.1.0
hyperlink==21.0.0
icecream==2.1.0
idna==2.10
incremental==17.5.0
itsdangerous==1.1.0
Jinja2==2.11.3
kiwisolver==1.3.1
MarkupSafe==1.1.1
matplotlib==3.3.4
msgpack==1.0.2
numpy==1.20.1
pandas==1.2.2
Pillow==8.1.0
plotly==4.14.3
psycopg2==2.8.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.6.0
pycparser==2.20
pydantic==1.7.3
Pygments==2.7.4
PyHamcrest==2.0.2
pyOpenSSL==20.0.1
pyparsing==2.4.7
python-dateutil==2.8.1
python-dotenv==0.15.0
python-multipart==0.0.5
pytz==2021.1
PyYAML==5.4.1
redis==3.5.3
requests==2.25.1
retrying==1.3.3
selenium==3.141.0
service-identity==18.1.0
six==1.15.0
sqlparse==0.4.1
starlette==0.13.6
toml==0.10.2
Twisted==20.3.0
txaio==20.12.1
typing-extensions==3.7.4.3
ujson==4.0.2
urllib3==1.26.3
uvicorn==0.13.3
watchgod==0.6
websockets==8.1
Werkzeug==1.0.1
zope.interface==5.2.0
